Refactored Stars out of containing module
