+++ client/components/ReviewListItem.jsx
@@ -1,11 +1,13 @@
 import React from 'react';
+import Stars from './Stars';
 
 const ReviewListItem = (props) => {
   const classes = { ...props };
   return (
     <div>
       <div id="review">
-        <h1>Rating: {classes.review.rating} {classes.review.name} - {classes.review.createdAt}</h1>
+        <Stars average={classes.review.rating} />
+        <h1>{classes.review.name} - {classes.review.createdAt}</h1>
         <h2>{classes.review.title}</h2>
         <h3>{classes.review.text}</h3>
         <h3> {classes.review.recommend} I recommend</h3>
